Reorganizing some aspects of the operational NWP turbulence scheme - Part 2: Modularizing the basic initialization of turbulence for NWP
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1781

### Short description

Initialization of the turbulence-model for NWP (contained in the modules 'turb_[transfer|diffusion|utilities]' is gained by some special treatment within the usual time-step call of the SUBs 'turb_[tran|diff]', since those calculations, which are only necessary for initialization, require a couple of general preparations intrinsically tied to internal formulations of that turbulence model.
Therefore, extracting those calculations in terms of a separate upstream initialization routine, would cause a rather cumbersome/intricate overcomplication and bloating of the code.

Nevertheless, the main part of particular initialization-calculations is about a simplified equilibrium-solution of the TKE-equation, which is (so far) coded several times in the SUBs 'turb_[tran|diff]' in only slightly modified modes. 

In the sense of a stronger modularization and tightening of the code, these similar calculations are now unified in the additional SUB 'init_basic_atmo_turb' of module 'turb_utilities', which is called in different modes.

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>