[nwp] Crosscheck for SB 2mom with parameterized deep convection
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1783

The combination of SB two-moment, inwp_gscp=4/5/6/7, with parameterized deep convection has not been tested, is not part of builtbot, and is not supported by DWD. Without the crosscheck there might be the false impression that this is a tested and supported configuration. This would be wrong and misleading. Conceptually there are many issues when running the SB two-moment scheme globally and/or in combination with parameterized deep convection. The users can always remove the CALL finish and proceed at their own risk.

Approved-by: Alberto de Lozar <alberto.lozar-de@dwd.de>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>