Bug fix for nudging interpolation
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1795

### Short description

Following an earlier modification of the subroutine `interpol_scal_grf` for lateral boundary interpolation, the interface of `interpol_scal_nudging` is extended by an additional argument providing the second dimension of a local auxiliary field for the case that it differs from the number of model levels. This is relevant if buffer arrays collecting a number of 2D variables are passed to `interpol_scal_nudging`, such as for nest initialization during runtime. 

Up to now, the maximum size of such buffer arrays is set manually in `interpol_scal_nudging`, but this has proven to be error-prone. In fact, adjusting this size was forgotten during the last extension of the nest initialization routine, causing an array bound violation for configurations with few vertical levels. This is now fixed in a sustainable way.

Co-authored-by: Guenther Zaengl <Guenther.Zaengl@dwd.de>
Approved-by: Daniel Rieger <daniel.rieger@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>