Lhn event bugfix
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1792

A bug in the LHN event handling is fixed. The event control was incorrectly computed for child domains, leading to an alternating on-off-behavior at the first nest level and (due to a second bug), to an invalid memory access at the second nest level of a three-domain configuration like ICON-D05.

Co-authored-by: Guenther Zaengl <Guenther.Zaengl@dwd.de>
Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>