New deep convection options to improve precipitation in the tropics
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1778

## Feature Description
<!--Describe your feature and important technical aspects-->

This merge request adds new options for the Tiedtke-Bechtold convection scheme of ICON. To have a smoother transition for land to ocean, and improve the precipitation over tropical islands, an interpolation using cloud droplet number can replace the land-sea mask. For the parcel ascent the original IFS implementation with the cubic profile is restored as an option. A RH-dependent detrainment formula can be used via namelist parameter. The fractional mass flux of the downdraft over land and ocean, which are important parameters for the precipitation efficiency of convective systems, are now in the namelist as well. These changes aim to improve precipitation in the tropics especially in the climate mode.

Co-authored-by: Guenther Zaengl <Guenther.Zaengl@dwd.de>
Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Approved-by: Martin Köhler <martin.koehler@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>