[nwp] Add CI scripts for Euler machine
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1796

### Short description

Add CI scripts for Euler (ETHZ) machine

### Detailed description

This merge request adds scripts to support both local and automated testing of ICON on the ETHZ Euler cluster. The scripts are placed in `scripts/jenkins_scripts`, which - as the name indicates - are independent of the official Buildbot-based testing.

The motivation is to enable testing of official ICON releases on Euler without requiring a custom repository or additional patching. This setup makes it easier to run CI-style tests on Euler in a maintainable and reproducible way.

---

Co-authored-by: Michael Jähn <mjaehn@ethz.ch>
Approved-by: Annika Lauber <annika.lauber@c2sm.ethz.ch>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>