More flexible options for entrainment/detrainment tuning in convection scheme
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1810

Building upon !1778, it turned out that the tuning option for the detrainment profile should be extended to allow for a RH-prefactor differing from 1 in order to obtain optimal results. In addition, an analogous tuning option is implemented for the RH-dependence of entrainment. Note that the latter does not affect the test parcel ascent because this would introduce unwanted side effects of the precipitation intensity spectrum (drizzle bias).

Co-authored-by: Guenther Zaengl <Guenther.Zaengl@dwd.de>
Approved-by: Axel Seifert <axel.seifert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>