[nwp] Move sea ice/snow albedo parameters to the namelist
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1775

### Short description

[nwp] move sea ice/snow albedo parameters to namelist

### Detailed description

It is necessary to move the sea ice and snow albedo parameters to the namelsit as these parameters may need to be tuned in ICON-XPP for the CMIP7 simulations. Presently, the below formula is used for calculating the equilibrium sea ice albedo:

alb_seaice_equil = csalb(ist_seaice) * ( 1.0_wp - 0.3143_wp * EXP(-0.35_wp*(tf_fresh-t_ice)) )

In which a derived constant 0.3143 is equal to (albsi_max-albsi_min)/albsi_max, where albsi_max=csalb(ist_seaice)=0.7 and albsi_min=0.48.

With proposed code changes, this formula is replaced with below format using two parameters (albsi_max and albsi_min) that are added to the namelist:

alb_seaice_equil =  albsi_max - (albsi_max-albsi_min)*EXP(-0.35*(tf_fresh-t_ice))

The parameters for calculating the snow albedo on sea ice (albsi_snow_max and albsi_snow_min) are hard-wired in the model. With proposed changes, these will also be moved to the namelist.

In the namelist, these four parameters retain the default values in the model. However, there are small differences in results impacted by the change. The differences may arise from the rounding of these numbers.

Three initial tests were performed (with the PI control set up) and the respective pyicon plots are linked below:

1) 11-year simulation with the orginal code (icon-nwp master)
https://swift.dkrz.de/v1/dkrz_f7eaee2e-e347-429e-b378-a6f4c4e455c4/GZENG/pyicon_levante/qp-slo1845_nwp/tave_1851-01-01-1862-01-01/qp_index.html

2) 11-year simulation with code changes (icon-nwp with branch "seaice-albedo-namelist", and the default values of these four parameters)
https://swift.dkrz.de/v1/dkrz_f7eaee2e-e347-429e-b378-a6f4c4e455c4/GZENG/pyicon_levante/qp-slo1845_seaice/tave_1851-01-01-1862-01-01/qp_index.html

3) 11-year simulation with code changes (as above, but set albsi_min = 0.47999 instead of 0.48 to make  (albsi_max-albsi_min)/albsi_max = 0.3143 as close as possible. 
https://swift.dkrz.de/v1/dkrz_f7eaee2e-e347-429e-b378-a6f4c4e455c4/GZENG/pyicon_levante/qp-slo1845_seaice2/tave_1851-01-01-1862-01-01/qp_index.html
---

- [x] I have read the [contribution guidelines for merge requests](/CONTRIBUTING.md#merge-requests) (last update: 2025-03-26).
- [x] I have considered updating the [docs.icon-model.org documentation](https://gitlab.dkrz.de/icon/wiki/-/wikis/Documentation).

Co-authored-by: Daniel Reinert <daniel.reinert@dwd.de>
Co-authored-by: Roland Wirth <roland.wirth@dwd.de>
Co-authored-by: dmitrii.mironov <dmitrii.mironov@dwd.de>
Approved-by: Roland Wirth <roland.wirth@dwd.de>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>